Merged raw Enrollment Data CSV Files, Updated the README.MD
Used Microsoft 365, and Power Query Functionality in Excel to merge 110 CSV Files with Historic Enrollment Status Data.

NOTE:  The Combined CSV File is stored in the GitHub, to maintain the immutable status of the raw data. DO NOT MESS WITH THE RAW DATA. CREATE A PYTHON NOTEBOOK TO PROCESS THIS RAW DATA, AND STORE IT IN THE 02_processed FOLDER.